+++ backend/models/user.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger
 import datetime
 from .base import Base
 
 class User(Base):
     __tablename__ = 'users'
     
-    telegram_id = Column(Integer, primary_key=True, index=True)
+    telegram_id = Column(BigInteger, primary_key=True, index=True)
     first_name = Column(String, nullable=True)
     username = Column(String, nullable=True)
     phone = Column(String, nullable=True)
